Clarification du code Délétion des lignes de code inutiles où à retravailler plus tard L'ajout des nodes se fait maintenant au centre de l'éditeur de nodes

+++ Nodes/NodesAssembly.cs
@@ -35,7 +35,7 @@ namespace SkyView.Nodes {
         private long _IdSelected;
         private Size _ProjectSize;
 
-        public void AddNode(NodeType type, int x, int y) {
+        public void AddNode(NodeType type, double x, double y) {
             LogicalNode newLogicalNode = new LogicalNode(type, x, y);
             NodesCollection.Add(newLogicalNode);
             IdSelected = newLogicalNode.Id;
